TSAGE: Fix compilation on Mac OS X PPC.
Mac OS X PPC's standard C(++) library does not feature any sqrtf function,
since the parameter is double anyway I changed it to use sqrt for now.

We might want to consider adding some global macros for that on such
platforms. Another instance where the same issues caused problems in the past
is the MT-32 emulator. See audio/softsynth/mt32/synth.cpp lines 30 to 43 for
reference.

+++ engines/tsage/core.cpp
@@ -925,7 +925,7 @@ int PlayerMover::findDistance(const Common::Point &pt1, const Common::Point &pt2
 	double yy = diff * 8.0 / 7.0;
 	yy *= yy;
 
-	return (int)sqrtf(xx + yy);
+	return (int)sqrt(xx + yy);
 }
 
 bool PlayerMover::sub_F8E5(const Common::Point &pt1, const Common::Point &pt2, const Common::Point &pt3,
